Fixed color variables so that they will be displayed clearly instead of disappearing into the backlight color.
The variables RED, GREEN and BLUE were set to [1,0,0] etc. This meant that when the example is run the displayed text is the same color than the backlight text making it look like it is not displayed at all. By changing the variables to a [100, 0,0] format the text becomes clearly visible. This makes the example work out of the box for new users.